feat: add controler functionality to change books shelf (2)
1)Add changeBookShelf() method in BookApp component. Method
receives the book user selected and the input of the new shelf
the user wants to move the book. setState() is used to change
the shelf property of the selected book in previous state, and
update the state with the new value.
2)Add changeBookShelf() as a prop in BookShelf component.
3)Add changeBookShelf() as a prop in Book component.
4)Go in Book.js and add a onChange event in select tag,
that uses the changeBookShelf() method. Select has a value
attribute that sets it's default value as the current shelf of
the book.